requirements.txt
python-jose==3.3.0
requests==2.28.2
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
python-dotenv==1.0.0
jwt==1.3.1
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-instrumentation-wsgi==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
